bitirme projesi/requirements_inference.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
torch>=2.3.0
torchvision>=0.18.0
timm>=1.0.9
pillow>=10.3.0
python-multipart>=0.0.9
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
segment-anything>=1.0
